requirements.txt
Django==1.9.4
django-filter==0.13.0
django-js-reverse==0.7.0
django-webpack-loader==0.3.0
djangorestframework==3.3.3
Markdown==2.6.6
